pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "clarke"
version = "0.1.0"
description = "Cognitive Learning Augmentation Retrieval Knowledge Engine"
requires-python = ">=3.12"
dependencies = [
    # API
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    # Orchestration
    "langgraph>=0.2",
    "langchain-core>=0.3",
    # Model gateway
    "litellm>=1.40",
    # Database
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    # Telemetry
    "opentelemetry-api>=1.25",
    "opentelemetry-sdk>=1.25",
    "opentelemetry-instrumentation-fastapi>=0.46b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.46b0",
    "opentelemetry-exporter-otlp>=1.25",
    # Retrieval
    "qdrant-client>=1.9",
    "tiktoken>=0.7",
    "numpy>=1.26",
    # Clustering
    "scikit-learn>=1.3",
    # Graph
    "neo4j>=5.0",
    # Background jobs
    "temporalio>=1.5",
    # Utilities
    "structlog>=24.1",
    "httpx>=0.27",
    "python-ulid>=2.7",
]

[project.optional-dependencies]
rerank = [
    "sentence-transformers>=3.0",
]
eval = [
    "arize-phoenix>=4.0",
]
docs = [
    "unstructured[all-docs]>=0.15",
]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.5",
    "mypy>=1.10",
    "pre-commit>=3.7",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "litellm.*",
    "langgraph.*",
    "langchain_core.*",
    "opentelemetry.*",
    "qdrant_client.*",
    "sentence_transformers.*",
    "neo4j.*",
    "sklearn.*",
    "phoenix.*",
    "arize.*",
    "temporalio.*",
]
ignore_missing_imports = true
